add the SVG files for the Underbar logo
I came up with the icon idea in a private conversation with Thibault
Duponchelle on IRC, on 2025-01-12. I produced the design on 2025-01-13
using Inkscape, and refined it a little before committing.

The font used for the dollar sign ($) and underscore (_) characters is
Courier New Bold. The "beer mug" (U+1F37A) comes from OpenMoji.